requirements.txt
Django==3.0.9
mysqlclient==1.4.6
gunicorn==20.0.4
django-rosetta==0.9.4
django-currentuser==0.5.0
djangorestframework==3.11.0
